server/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
torch==2.1.0
torchvision==0.16.0
opencv-python-headless==4.7.0.72
numpy==1.24.3
Pillow==9.5.0
pydantic==1.10.9
nest-asyncio==1.5.6
python-multipart==0.0.5
pandas==2.0.0
